% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie.r
\name{CIE}
\alias{CIE}
\title{Constructor of CIE weighted waveband}
\usage{
CIE(norm = 298, w.low = 250, w.high = 400)
}
\arguments{
\item{norm}{normalization wavelength (nm)}

\item{w.low}{short-end boundary wavelength (nm)}

\item{w.high}{long-end boundary wavelength (nm)}
}
\value{
a waveband object wavelength defining wavelength range, weighting
  function and normalization wavelength.
}
\description{
Erythema BSWF
}
\examples{
CIE()
CIE(300)

}
\references{
Webb, A. (20XX)
}
\seealso{
\code{\link[photobiology]{waveband}}

Other BSWF weighted wavebands: 
\code{\link{CH4}()},
\code{\link{DNA_GM}()},
\code{\link{DNA_N}()},
\code{\link{DNA_P}()},
\code{\link{FLAV}()},
\code{\link{GEN_G}()},
\code{\link{GEN_M}()},
\code{\link{GEN_T}()},
\code{\link{ICNIRP}()},
\code{\link{PG}()}
}
\concept{BSWF weighted wavebands}
