% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.data.r
\docType{data}
\name{CIE1924_lef.spct}
\alias{CIE1924_lef.spct}
\title{CIE1924 luminous efficiency function (photopic human vision)}
\format{
A response.spct object with 471 rows and 2 variables
}
\description{
A dataset containing the wavelengths at a 1 nm interval.
Tabulated values for quantum luminous efficiency according to CIE1924.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item s.q.response
}
}
\note{
This luminous efficiency function understimates the renponse to short
  wavelengths.
}
\references{
\url{http://www.cvrl.org/} downloaded on 2015-01-24
}
\keyword{datasets}
