% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.data.r
\docType{data}
\name{CIE1951_scotopic_lef.spct}
\alias{CIE1951_scotopic_lef.spct}
\title{Luminous efficiency function (scotopic human vision)}
\format{
A response.spct object with 401 rows and 2 variables
}
\description{
A dataset containing the wavelengths at a 1 nm interval. Tabulated values for
quantum luminous efficiency at low light levels according to CIE1951.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item s.q.response
}
}
\references{
\url{http://www.cvrl.org/} downloaded on 2015-01-24
}
\keyword{datasets}
