% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.data.r
\docType{data}
\name{CIE2008_lef2deg.spct}
\alias{CIE2008_lef2deg.spct}
\title{CIE2008 luminous efficiency function (2-deg) (photopic human vision)}
\format{
A response.spct object with 441 rows and 2 variables
}
\description{
A dataset containing the wavelengths at a 1 nm interval.
Tabulated values for quantum luminous efficiency according to CIE2008 for 2 degrees.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item s.q.response
}
}
\references{
\url{http://www.cvrl.org/} downloaded on 2015-01-24
}
\keyword{datasets}
