% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.gm.r
\name{DNA.GM}
\alias{DNA.GM}
\title{Constructor of DNA damage (SETLOW) weighted waveband}
\usage{
DNA.GM(norm = 300, w.low = 275, w.high = 400)
}
\description{
Naked DNA damage BSWF, Green and Miller's formulation.
}
\seealso{
\code{\link{DNA_GM}}
}
\keyword{internal}
