% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.n.r
\name{DNA.N}
\alias{DNA.N}
\title{Constructor of DNA damage (SETLOW) weighted waveband}
\usage{
DNA.N(norm = 300, w.low = 275, w.high = 400)
}
\description{
Naked DNA damage BSWF
}
\seealso{
\code{\link{DNA_N}}
}
\keyword{internal}
