% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.p.r
\name{DNA.P}
\alias{DNA.P}
\title{Constructor of DNA damage (Quaite) weighted waveband}
\usage{
DNA.P(norm = 300, w.low = 275, w.high = 400)
}
\description{
Plant DNA damage BSWF as formulated by Musil.
}
\seealso{
\code{\link{DNA_P}}
}
\keyword{internal}
