% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.gm.r
\name{DNA_GM}
\alias{DNA_GM}
\title{Constructor of DNA damage (SETLOW) weighted waveband}
\usage{
DNA_GM(norm = 300, w.low = 275, w.high = 400)
}
\arguments{
\item{norm}{normalization wavelength (nm)}

\item{w.low}{short-end boundary wavelength (nm)}

\item{w.high}{long-end boundary wavelength (nm)}
}
\value{
a waveband object wavelength defining wavelength range, weighting function
and normalization wavelength.
}
\description{
Naked DNA damage BSWF, Green and Miller's formulation.
}
\examples{
DNA_GM()
DNA_GM(300)

}
\seealso{
\code{\link[photobiology]{waveband}}

Other BSWF weighted wavebands:
\code{\link[=CH4]{CH4()}},
\code{\link[=DNA_N]{DNA_N()}},
\code{\link[=DNA_P]{DNA_P()}},
\code{\link[=FLAV]{FLAV()}},
\code{\link[=GEN_G]{GEN_G()}},
\code{\link[=GEN_M]{GEN_M()}},
\code{\link[=GEN_T]{GEN_T()}},
\code{\link[=PAR]{PAR()}},
\code{\link[=PG]{PG()}},
\code{\link[=UV_health_hazard]{UV_health_hazard()}},
\code{\link[=erythema]{erythema()}}
}
\concept{BSWF weighted wavebands}
