% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.g.r
\name{GEN.G}
\alias{GEN.G}
\title{Constructor of GPAS (Green) weighted waveband}
\usage{
GEN.G(norm = 300, w.low = 275, w.high = 313.3)
}
\description{
Generalized Plant Action BSWF of Caldwell as formulated by Green et al.
}
\seealso{
\code{\link{GEN_G}}
}
\keyword{internal}
