% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.m.r
\name{GEN.M}
\alias{GEN.M}
\title{Constructor of GPAS (Micheletti) weighted waveband}
\usage{
GEN.M(norm = 300, w.low = 275, w.high = 313.3)
}
\description{
Generalized Plant Action BSWF of Caldwell as formulated by Micheletti et al.
}
\seealso{
\code{\link{GEN_M}}
}
\keyword{internal}
