% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.t.r
\name{GEN.T}
\alias{GEN.T}
\title{Constructor of GPAS (Timijan) weighted waveband}
\usage{
GEN.T(norm = 300, w.low = 275, w.high = 345)
}
\description{
Generalized Plant Action BSWF of Caldwell as formulated by Timijan et al.
}
\seealso{
\code{\link{GEN_T}}
}
\keyword{internal}
