% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat.bands.r
\name{Landsat_bands}
\alias{Landsat_bands}
\alias{RBV_bands}
\alias{MSS_bands}
\alias{OLI_bands}
\alias{TIRS_bands}
\alias{ETM_bands}
\title{Constructor of lists of wavebands matching Landsat imagers}
\usage{
Landsat_bands(std = "L8")

RBV_bands(std = "LandsatRBV")

MSS_bands(std = "LandsatMSS")

OLI_bands(std = "LandsatOLI")

TIRS_bands(std = "LandsatTIRS")

ETM_bands(std = "LandsatETM")
}
\arguments{
\item{std}{a character string "L1"..."L9", for missions, "LandsarRBV",
"LandsatMSS", etc. for imagers.}
}
\value{
a list of wavebands
}
\description{
Lists of \code{waveband} objects describing bands defined as ranges of
wavelengths according to NASA and USGS manuals for different Landsat
missions.
}
\details{
Over the many years of the different Landsat missions, instruments
  changed and consequently also some of the wavelength ranges captured in the
  images. The definitions are as \emph{rectangular} windows, while the true
  response functions deviate to some extent from these ideal definitions.

  The individual \code{waveband} objects included in the returned \code{list}
  objects are constructed with functions \code{\link{Purple}()},
  \code{\link{Blue}()}, \code{\link{Green}()}, \code{\link{Red}()},
  \code{\link{Far_red}()}, \code{\link{NIR}()}, \code{\link{SWIR1}()}, and
  \code{\link{SWIR2}()}.

  See \url{https://landsat.usgs.gov/spectral-characteristics-viewer} for
  detailed sensitivity spectra for the different bands of the imaginers.
}
\examples{
Landsat_bands("L1")
Landsat_bands("L8")
OLI_bands()
TIRS_bands()

}
\seealso{
\code{\link[photobiology]{waveband}()} for a description of the
  class and its constructors.

Other lists of unweighted wavebands:
\code{\link[=IR_bands]{IR_bands()}},
\code{\link[=Plant_bands]{Plant_bands()}},
\code{\link[=UV_bands]{UV_bands()}},
\code{\link[=VIS_bands]{VIS_bands()}}
}
\concept{lists of unweighted wavebands}
