% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.q.fun.r
\name{PG_q_fun}
\alias{PG_q_fun}
\title{Gives values for the Plant Growth BSWF as a function of wavelength}
\usage{
PG_q_fun(w.length)
}
\arguments{
\item{w.length}{numeric array of wavelengths (nm)}
}
\value{
a numeric array of the same length as \code{w.length} with values for
  the BSWF normalized as in the original source (300 nm)
}
\description{
This function gives a set of numeric multipliers that can be used
as a weight to calculate effective doses and irradiances. The
returned values are on quantum based effectiveness relative units.
}
\note{
We follow the original defition here for the equation, with no
limtation to the wavelength range. However, be aware that in practice
it is not used for long wavelengths (different limits between 366 nm and
400 nm have been used by different authors).
}
\examples{
PG_q_fun(293:400)

}
\seealso{
Other BSWF functions:
\code{\link[=CH4_e_fun]{CH4_e_fun()}},
\code{\link[=CH4_q_fun]{CH4_q_fun()}},
\code{\link[=CIE_e_fun]{CIE_e_fun()}},
\code{\link[=CIE_q_fun]{CIE_q_fun()}},
\code{\link[=DNA_GM_q_fun]{DNA_GM_q_fun()}},
\code{\link[=DNA_P_q_fun]{DNA_P_q_fun()}},
\code{\link[=FLAV_q_fun]{FLAV_q_fun()}},
\code{\link[=GEN_G_q_fun]{GEN_G_q_fun()}},
\code{\link[=GEN_M_q_fun]{GEN_M_q_fun()}},
\code{\link[=GEN_T_q_fun]{GEN_T_q_fun()}},
\code{\link[=ICNIRP_e_fun]{ICNIRP_e_fun()}}
}
\concept{BSWF functions}
