% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.data.r
\name{photopic_sensitivity}
\alias{photopic_sensitivity}
\title{Photopic sensitivity of the human eye}
\format{
A single numeric value
}
\usage{
photopic_sensitivity
}
\description{
Constant value used in the definition of Lumen
1 W is equal to 683 Lumen at 555 nm
}
\details{
A single numeric value
}
