% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.data.r
\name{scotopic_sensitivity}
\alias{scotopic_sensitivity}
\title{Scotopic sensitivity of the human eye}
\format{
A single numeric value
}
\usage{
scotopic_sensitivity
}
\description{
Constant value for human vision under very weak illumination
1 W is equal to 1699 Lumen at 507 nm
}
\details{
A single numeric value
}
