% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel2.bands.r
\name{Sentinel2_bands}
\alias{Sentinel2_bands}
\title{Constructor of lists of wavebands matching Sentinel-2 multi spectral
imager (MSI).}
\usage{
Sentinel2_bands(std = "MSI")
}
\arguments{
\item{std}{a character string "MSI", for the imager, or "10m", "20m", or
"60m" based on image spatial resolution.}
}
\value{
a list of wavebands
}
\description{
Defined as ranges of wavelengths according to descriptions in the Sentinel-2
MSI manual from ESA. The definitions are as \emph{rectangular} windows, while
the true response functions deviate to some extent from these ideal
definitions.
}
\details{
The ground-level spatial resolution of the images depends on the
  wavelength, grouped into three resolutions of 10, 20 and 60 m. The
  wavelength ranges implemented are the nominal ones which very slightly
  differ from the actual ones and between Sentinel-2a and Sentinel-2b
  (see \url{https://www.earthdata.nasa.gov/data/instruments/sentinel-2-msi}).

  The individual \code{waveband} objects included in the returned \code{list}
  objects are constructed with functions \code{\link{Blue}()},
  \code{\link{Green}()}, \code{\link{Red}()}, \code{\link{Far_red}()}, and
  \code{\link{NIR}()}.
}
\examples{
Sentinel2_bands()
Sentinel2_bands("10m")

}
\references{
SUHET (2015) SENTINEL-2 User Handbook, ESA Standard Document Date 24/07/2015
Issue 1 Rev 2. European Space Agency.
}
\seealso{
\code{\link[photobiology]{waveband}()} for a description of the
  class and its constructors.
}
\concept{lists of \code{waveband} objects defining wavelength ranges, i.e.,
  lacking weighting functions.}
