% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.data.r
\docType{data}
\name{SetlowTUV.spct}
\alias{SetlowTUV.spct}
\title{Setlow's action spectrum for DNA damage}
\format{
A response.spct object with 1082 rows and 2 variables
}
\description{
A dataset containing the wavelengths at a 0.1 nm interval. Tabulated values
for Setlow's naked DNA damage action spectrum as used in the TUV model.
}
\details{
The variables are as follows:

\itemize{
  \item w.length (nm)
  \item s.e.response
}
}
\references{
\url{http://uv.biospherical.com/Version2/description-Version2-Database3.html} downloaded 2015-02-07
}
\keyword{datasets}
