% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{VIS}
\alias{VIS}
\title{Constructor of VIS waveband}
\usage{
VIS(std = "ISO")
}
\arguments{
\item{std}{a character string "ISO" or "RS" (remote sensing).}
}
\value{
A waveband object wavelength defining a wavelength range.
}
\description{
Visible (to humans) radiation (380...760 nm) according to ISO standard
definition, no weighting applied. For \code{std = "RS"} the returned range is
the same as for \code{PAR()}. The panchromatic bands of Landsat missions are
also supported.
}
\examples{
VIS()
VIS("ISO")
VIS("LandsatOLI")
VIS("Landsat7")
VIS("Pan.RBV.Landsat3")

}
\seealso{
\code{\link[photobiology]{waveband}}

Other unweighted wavebands:
\code{\link[=Blue]{Blue()}},
\code{\link[=Far_red]{Far_red()}},
\code{\link[=Green]{Green()}},
\code{\link[=IR]{IR()}},
\code{\link[=Orange]{Orange()}},
\code{\link[=Purple]{Purple()}},
\code{\link[=Red]{Red()}},
\code{\link[=UV]{UV()}},
\code{\link[=UVA]{UVA()}},
\code{\link[=UVB]{UVB()}},
\code{\link[=UVC]{UVC()}},
\code{\link[=Yellow]{Yellow()}}
}
\concept{unweighted wavebands}
