% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.bands.r
\name{VIS_bands}
\alias{VIS_bands}
\title{Constructor of lists of VIS wavebands}
\usage{
VIS_bands(std = "ISO")
}
\arguments{
\item{std}{a character string "ISO".}
}
\value{
a list of wavebands
}
\description{
Defined according to "ISO".
}
\examples{
VIS_bands()
VIS_bands("ISO")

}
\seealso{
\code{\link[photobiology]{waveband}}, \code{\link{Purple}},
  \code{\link{Blue}}, \code{\link{Green}}, \code{\link{Yellow}},
  \code{\link{Orange}}, and \code{\link{Red}}.

Other lists of unweighted wavebands:
\code{\link[=IR_bands]{IR_bands()}},
\code{\link[=Landsat_bands]{Landsat_bands()}},
\code{\link[=Plant_bands]{Plant_bands()}},
\code{\link[=UV_bands]{UV_bands()}}
}
\concept{lists of unweighted wavebands}
