% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue.r
\name{Blue}
\alias{Blue}
\title{Constructor of blue waveband}
\usage{
Blue(std = "ISO")
}
\arguments{
\item{std}{a character string "ISO", "Sellaro" (plant biology), or "RS"
(remote sensing), or Landsat imagers, "LandsatTM", "LandsatETM", or
"LandsatOLI".}
}
\value{
a waveband object wavelength defining a wavelength range.
}
\description{
Default defined according to "ISO".
}
\examples{
Blue()
Blue("ISO")
Blue("Sellaro")

}
\seealso{
\code{\link{new_waveband}}  \code{\link{waveband}}

Other unweighted wavebands: 
\code{\link{Far_red}()},
\code{\link{Green}()},
\code{\link{IR}()},
\code{\link{Orange}()},
\code{\link{Purple}()},
\code{\link{Red}()},
\code{\link{UVA}()},
\code{\link{UVB}()},
\code{\link{UVC}()},
\code{\link{UV}()},
\code{\link{VIS}()},
\code{\link{Yellow}()}
}
\concept{unweighted wavebands}
