% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue.r
\name{Blue}
\alias{Blue}
\title{Definition of list of VIS wavebands}
\usage{
Blue(std = "ISO")
}
\arguments{
\item{std}{a character string "ISO", or "Sellaro"}
}
\value{
a waveband object wavelength defining a wavelength range.
}
\description{
Defined according to "ISO".
}
\examples{
Blue()
Blue("ISO")
Blue("Sellaro")

}
\seealso{
\code{\link{new_waveband}}  \code{\link{waveband}}

Other unweighted wavebands: \code{\link{Far_red}},
  \code{\link{Green}}, \code{\link{Orange}},
  \code{\link{Purple}}, \code{\link{Red}},
  \code{\link{UVA}}, \code{\link{UVB}}, \code{\link{UVC}},
  \code{\link{UV}}, \code{\link{VIS}}, \code{\link{Yellow}}
}

