% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth.r
\name{CH4}
\alias{CH4}
\title{Constructor of CH4 production from pectin weighted waveband}
\usage{
CH4(norm = 300, w.low = 275, w.high = 400)
}
\arguments{
\item{norm}{normalization wavelength (nm)}

\item{w.low}{short-end boundary wavelength (nm)}

\item{w.high}{long-end boundary wavelength (nm)}
}
\value{
a waveband object wavelength defining wavelength range, weighting function
and normalization wavelength.
}
\description{
Methane production from pectin BSWF
}
\examples{
CIE()
CIE(300)

}
\seealso{
\code{\link[photobiology]{waveband}}

Other BSWF weighted wavebands: 
\code{\link{CIE}()},
\code{\link{DNA_GM}()},
\code{\link{DNA_N}()},
\code{\link{DNA_P}()},
\code{\link{FLAV}()},
\code{\link{GEN_G}()},
\code{\link{GEN_M}()},
\code{\link{GEN_T}()},
\code{\link{ICNIRP}()},
\code{\link{PG}()}
}
\concept{BSWF weighted wavebands}
