% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth.e.fun.r
\name{CH4_e_fun}
\alias{CH4_e_fun}
\title{Gives values for the CH4 production from pectin BSWF as a function of wavelength}
\usage{
CH4_e_fun(w.length)
}
\arguments{
\item{w.length}{numeric array of wavelengths (nm)}
}
\value{
a numeric array of the same length as \code{w.length} with values for the BSWF normalized
as in the original source (300 nm) and based on energy effectiveness.
}
\description{
This function gives a set of numeric multipliers that can be used
as a weight to calculate effective doses and irradiances. The
returned values are on quantum based effectiveness relative units.
}
\examples{
CH4_e_fun(293:400)

}
\references{
Bloom, A. A.; Lee-Taylor, J.; Madronich, S.; Messenger, D. J.; Palmer, P. I.;
Reay, D. S. & McLeod, A. R. (2010) Global methane emission estimates from
ultraviolet irradiation of terrestrial plant foliage. New Phytologist,
Blackwell Publishing Ltd, 187, 417--425 .
}
\seealso{
Other BSWF functions: 
\code{\link{CH4_q_fun}()},
\code{\link{CIE_e_fun}()},
\code{\link{CIE_q_fun}()},
\code{\link{DNA_GM_q_fun}()},
\code{\link{DNA_P_q_fun}()},
\code{\link{FLAV_q_fun}()},
\code{\link{GEN_G_q_fun}()},
\code{\link{GEN_M_q_fun}()},
\code{\link{GEN_T_q_fun}()},
\code{\link{ICNIRP_e_fun}()},
\code{\link{PG_q_fun}()}
}
\concept{BSWF functions}
