% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green.r
\name{Green}
\alias{Green}
\title{Constructor of green waveband}
\usage{
Green(std = "ISO")
}
\arguments{
\item{std}{a character string "ISO", "Sellaro" or "LandsatRBV", and
equivalent names for Landsat imagers.}
}
\value{
a waveband object wavelength defining a wavelength range.
}
\description{
Green radiation according to ISO or as commonly defined in plant
photobiology, no weighting applied.
}
\note{
When released, this package will replace the package UVcalc.
}
\examples{
Green()
Green("ISO") # 500 to 570
Green("Sellaro") # 500 to 570 nm

}
\seealso{
\code{\link[photobiology]{waveband}}

Other unweighted wavebands: 
\code{\link{Blue}()},
\code{\link{Far_red}()},
\code{\link{IR}()},
\code{\link{Orange}()},
\code{\link{Purple}()},
\code{\link{Red}()},
\code{\link{UVA}()},
\code{\link{UVB}()},
\code{\link{UVC}()},
\code{\link{UV}()},
\code{\link{VIS}()},
\code{\link{Yellow}()}
}
\concept{unweighted wavebands}
