% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icnirp.e.fun.r
\name{ICNIRP_e_fun}
\alias{ICNIRP_e_fun}
\title{ICNIRP UV health hazard BSWF as a function of wavelength}
\usage{
ICNIRP_e_fun(w.length)
}
\arguments{
\item{w.length}{numeric array of wavelengths (nm)}
}
\value{
a numeric array of the same length as \code{w.length} with values for
  the BSWF normalized as in the original source (270 nm) and based on energy
  effectiveness.
}
\description{
This function returns a vector of numeric multipliers that can be used as a
weight to calculate effective doses and irradiances. The returned values are
on energy based effectiveness relative units. The BSWF is formulated for the
range 210 nm to 400 nm.
}
\examples{
ICNIRP_e_fun(210:400)

}
\seealso{
Other BSWF functions: 
\code{\link{CH4_e_fun}()},
\code{\link{CH4_q_fun}()},
\code{\link{CIE_e_fun}()},
\code{\link{CIE_q_fun}()},
\code{\link{DNA_GM_q_fun}()},
\code{\link{DNA_P_q_fun}()},
\code{\link{FLAV_q_fun}()},
\code{\link{GEN_G_q_fun}()},
\code{\link{GEN_M_q_fun}()},
\code{\link{GEN_T_q_fun}()},
\code{\link{PG_q_fun}()}
}
\concept{BSWF functions}
