% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orange.r
\name{Orange}
\alias{Orange}
\title{Constructor of orange waveband}
\usage{
Orange(std = "ISO")
}
\arguments{
\item{std}{a character string "ISO"}
}
\value{
a waveband object wavelength defining a wavelength range.
}
\description{
Orange radiation (591...610 nm), no weighting
applied.
}
\examples{
Orange()
Orange("ISO")

}
\seealso{
Other unweighted wavebands: 
\code{\link{Blue}()},
\code{\link{Far_red}()},
\code{\link{Green}()},
\code{\link{IR}()},
\code{\link{Purple}()},
\code{\link{Red}()},
\code{\link{UVA}()},
\code{\link{UVB}()},
\code{\link{UVC}()},
\code{\link{UV}()},
\code{\link{VIS}()},
\code{\link{Yellow}()}
}
\concept{unweighted wavebands}
