% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.r
\name{PAR}
\alias{PAR}
\title{Constructor of PAR waveband}
\usage{
PAR(std = "Plant")
}
\arguments{
\item{std}{a character string "Plant" or "McCree"}
}
\value{
a waveband object wavelength defining a wavelength range.
}
\description{
Photosythetically active radiation (400-700 nm), no weighting
applied.
}
\examples{
PAR()
PAR("Plant")
}
\references{
McCree, K. J. (1972) The action spectrum, absorptance and quantum
  yield of photosynthesis in crop plants. Agricultural Meteorology, 9,
  191-216
}
\seealso{
\code{\link[photobiology]{waveband}}
}
