% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.r
\name{Purple}
\alias{Purple}
\title{Constructor of purple waveband}
\usage{
Purple(std = "ISO")
}
\arguments{
\item{std}{a character string "ISO", or Landsat imager "LandsatOLI".}
}
\value{
A waveband object wavelength defining a wavelength range.
}
\description{
Purple radiation (360...450 nm), no weighting
applied.
}
\examples{
Purple()
Purple("ISO")

}
\seealso{
\code{\link{new_waveband}}  \code{\link{waveband}}

Other unweighted wavebands: 
\code{\link{Blue}()},
\code{\link{Far_red}()},
\code{\link{Green}()},
\code{\link{IR}()},
\code{\link{Orange}()},
\code{\link{Red}()},
\code{\link{UVA}()},
\code{\link{UVB}()},
\code{\link{UVC}()},
\code{\link{UV}()},
\code{\link{VIS}()},
\code{\link{Yellow}()}
}
\concept{unweighted wavebands}
