% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv.r
\name{UV}
\alias{UV}
\title{Constructor of UV waveband}
\usage{
UV(std = "ISO")
}
\arguments{
\item{std}{"ISO" or "CIE"}
}
\value{
a waveband object wavelength defining a wavelength range.
}
\description{
UV: 100--400 nm.
}
\examples{
UV()
UV("ISO")

}
\references{
ISO and CIE standards
}
\seealso{
\code{\link{new_waveband}}  \code{\link{waveband}}

Other unweighted wavebands: 
\code{\link{Blue}()},
\code{\link{Far_red}()},
\code{\link{Green}()},
\code{\link{IR}()},
\code{\link{Orange}()},
\code{\link{Purple}()},
\code{\link{Red}()},
\code{\link{UVA}()},
\code{\link{UVB}()},
\code{\link{UVC}()},
\code{\link{VIS}()},
\code{\link{Yellow}()}
}
\concept{unweighted wavebands}
