% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uva.r
\name{UVA}
\alias{UVA}
\title{Definition of UV-A waveband}
\usage{
UVA(std = "ISO")
}
\arguments{
\item{std}{a character string "CIE", "ISO" or "none"}
}
\value{
a waveband object wavelength defining a wavelength range.
}
\description{
UV-A according to CIE and ISO standrads: 315--400 nm.
UV-A according to common non-standard practice: 320--400 nm.
}
\examples{
UVA()
UVA("none")
UVA("ISO")
UVA("CIE")

}
\seealso{
\code{\link[photobiology]{waveband}}

Other unweighted wavebands: \code{\link{Blue}},
  \code{\link{Far_red}}, \code{\link{Green}},
  \code{\link{Orange}}, \code{\link{Purple}},
  \code{\link{Red}}, \code{\link{UVB}}, \code{\link{UVC}},
  \code{\link{UV}}, \code{\link{VIS}}, \code{\link{Yellow}}
}

