% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvb.r
\name{UVB}
\alias{UVB}
\title{Constructor of UV-B waveband}
\usage{
UVB(std = "ISO")
}
\arguments{
\item{std}{a character string "CIE", "ISO", "medical" or "none"}
}
\value{
a waveband object wavelength defining a wavelength range.
}
\description{
UV-B according to CIE and ISO standrads: 280--315 nm. UV-B according to
common non-standard practice: 280--320 nm. UV-B according to medical or
dermatological non-standard practice: 280--320 nm.
}
\note{
The non-standard definition of UV-B using 320 nm as limit
should not be used in any new publications or work as it deviates from
the internationally accepted standards and its use leads to confusion. Its
inclusion in this package is to allow calculations needed to compare
new results and methods against old publications.
}
\examples{
UVB()
UVB("ISO")
UVB("CIE")
UVB("none")
UVB("medical")

}
\seealso{
\code{\link[photobiology]{waveband}}

Other unweighted wavebands: 
\code{\link{Blue}()},
\code{\link{Far_red}()},
\code{\link{Green}()},
\code{\link{IR}()},
\code{\link{Orange}()},
\code{\link{Purple}()},
\code{\link{Red}()},
\code{\link{UVA}()},
\code{\link{UVC}()},
\code{\link{UV}()},
\code{\link{VIS}()},
\code{\link{Yellow}()}
}
\concept{unweighted wavebands}
