% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{VIS}
\alias{VIS}
\title{Definition of VIS waveband}
\usage{
VIS(std = "ISO")
}
\arguments{
\item{std}{a character string "ISO"}
}
\value{
A waveband object wavelength defining a wavelength range.
}
\description{
Visible (to humnas) radiation (380...760 nm), no weighting
applied.
}
\examples{
VIS()
VIS("ISO")

}
\seealso{
\code{\link[photobiology]{waveband}}

Other unweighted wavebands: \code{\link{Blue}},
  \code{\link{Far_red}}, \code{\link{Green}},
  \code{\link{Orange}}, \code{\link{Purple}},
  \code{\link{Red}}, \code{\link{UVA}}, \code{\link{UVB}},
  \code{\link{UVC}}, \code{\link{UV}}, \code{\link{Yellow}}
}

