% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.bands.r
\name{VIS_bands}
\alias{VIS_bands}
\title{Definition of list of VIS wavebands}
\usage{
VIS_bands(std = "ISO")
}
\arguments{
\item{std}{a character string "ISO" (ignored)}
}
\value{
a list of wavebands
}
\description{
Defined according to "ISO".
}
\examples{
VIS_bands()
VIS_bands("ISO")

}
\seealso{
Other lists of unweighted wavebands: \code{\link{Plant_bands}},
  \code{\link{UV_bands}}
}

