% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yellow.r
\name{Yellow}
\alias{Yellow}
\title{Constructor of yellow waveband}
\usage{
Yellow(std = "ISO")
}
\arguments{
\item{std}{a character string "ISO"}
}
\value{
a waveband object wavelength defining a wavelength range.
}
\description{
Yellow radiation (570...591 nm), no weighting applied.
}
\examples{
Yellow()
Yellow("ISO")

}
\seealso{
\code{\link[photobiology]{waveband}}

Other unweighted wavebands: 
\code{\link{Blue}()},
\code{\link{Far_red}()},
\code{\link{Green}()},
\code{\link{IR}()},
\code{\link{Orange}()},
\code{\link{Purple}()},
\code{\link{Red}()},
\code{\link{UVA}()},
\code{\link{UVB}()},
\code{\link{UVC}()},
\code{\link{UV}()},
\code{\link{VIS}()}
}
\concept{unweighted wavebands}
